Update link to a cdn copy of jQuery 1.8.2
Also, remove references to the the scripts related to the nist generic
exit script and its css, since the page is not longer hosted at nist.
Also remove the reference included for the us federal government
analytics.

It is doubtful that jQuery is required in these pages other than the
nist exist script, but just in case the jQuery reference was updated to
a cdn copy of the version in the original web site (1.8.2).